requirements.txt
ollama
opencv-python
Pillow
numpy
chromadb
sentence-transformers
pyyaml
tqdm
fastapi
uvicorn
python-multipart
